refactor: extract methods that are not API related to a new utility handler
