+++ src/services/GameService.ts
@@ -1,29 +1,16 @@
-import { GameRequestParams } from '@data/requests'
-import {
-	Developer,
-	Game,
-	GamesPlatform,
-	Genre,
-	Platform,
-	Publisher,
-	Recommended,
-	Screenshots,
-	Store,
-	Tag
-}                            from '@data/types'
-import { ApiService }        from '@services/index.ts'
-import { IGameService }      from '@services/interfaces'
-import { ParserUtils }       from '@src/utils'
+import { GameRequestParams }              from '@data/requests'
+import { Game, Recommended, Screenshots } from '@data/types'
+import { ApiService }                     from '@services/index.ts'
+import { IGameService }                   from '@services/interfaces'
+import { ParserUtils }                    from '@src/utils'
+import { TypeUtils }                      from '@utils/TypeUtils.ts'
 
 
 export const GameService: IGameService = {
 	getAll,
 	getById,
 	getRecommendations,
-	getScreenshots,
-	mapToGame,
-	mapToGamesPlatforms,
-	mapToStores
+	getScreenshots
 }
 
 async function getAll(params?: GameRequestParams): Promise<Game[]> {
@@ -32,14 +19,14 @@ async function getAll(params?: GameRequestParams): Promise<Game[]> {
 		params: params ? ParserUtils.mapToSnakeCase(params) : {}
 	})
 
-	return response.data?.results.map(mapToGame) ?? []
+	return response.data?.results.map(TypeUtils.mapToGame) ?? []
 }
 
 async function getById(id: number): Promise<Game> {
 	const gameUrl  = ApiService.createRouteUrl(`games/${id}`)
 	const response = await ApiService.gameApi.get(gameUrl)
 
-	return mapToGame(response.data)
+	return TypeUtils.mapToGame(response.data)
 }
 
 async function getRecommendations(): Promise<Recommended> {
@@ -52,59 +39,6 @@ async function getRecommendations(): Promise<Recommended> {
 		recentScreenshots: await getScreenshots(recentId),
 		dailyScreenshots : await getScreenshots(dailyId)
 	}
-}
-
-function mapToGame(data: any): Game {
-	const mappedData = ParserUtils.mapToCamelCase(data) as Game
-
-	if (mappedData.platforms !== undefined) {
-		mappedData.platforms
-			= mapToGamesPlatforms(mappedData.platforms)
-	}
-
-	if (mappedData.developers) {
-		mappedData.developers = mappedData.developers.map(d => ParserUtils
-		.mapToCamelCase(d) as Developer)
-	}
-
-	if (mappedData.publishers) {
-		mappedData.publishers = mappedData.publishers.map(p => ParserUtils
-		.mapToCamelCase(p) as Publisher)
-	}
-
-	if (mappedData.tags) {
-		mappedData.tags = mappedData.tags.map(t => ParserUtils
-		.mapToCamelCase(t) as Tag)
-	}
-
-	if (mappedData.genres) {
-		mappedData.genres = mappedData.genres.map(g => ParserUtils
-		.mapToCamelCase(g) as Genre)
-	}
-
-	if (mappedData.stores) {
-		mappedData.stores = mapToStores(mappedData.stores)
-	}
-
-	return mappedData
-}
-
-function mapToStores(stores: Store[]): Store[] {
-	return stores.map(store => {
-		store.store
-			= ParserUtils.mapToCamelCase(store.store) as typeof store.store
-
-		return store
-	})
-}
-
-function mapToGamesPlatforms(platforms: GamesPlatform[]): GamesPlatform[] {
-	return platforms.map(mainPlatform => {
-		mainPlatform.platform = ParserUtils
-		.mapToCamelCase(mainPlatform.platform) as Platform
-
-		return ParserUtils.mapToCamelCase(mainPlatform) as GamesPlatform
-	})
 }
 
 async function getScreenshots(id: number): Promise<Screenshots> {
